Offline support for google-signin.
This updates both google-signin-aware and google-signin components to
allow an offline parameter, which will be used to get an auth code that
can then be exchanged for offline credentials.

It is an attempt to reapply the changes from
github.com/GoogleWebComponents/google-signin/compare/api2.0-offline
after the refactoring of google-signin.